refactor: better organisation of folders
