+++ src/components/ProtectedRoute.tsx
@@ -1,7 +1,7 @@
 import React, { ReactNode } from 'react';
 import { Navigate, useLocation } from 'react-router-dom';
 import { Box, CircularProgress, Typography } from '@mui/material';
-import { useAuth } from '../hooks/useAuth';
+import { useAuth } from './providers/useAuth';
 
 interface ProtectedRouteProps {
   children: ReactNode;
