+++ src/components/providers/useAuth.ts
@@ -1,5 +1,5 @@
 import { useContext } from 'react';
-import { AuthContext, AuthContextType } from '../contexts/AuthContext';
+import { AuthContext, AuthContextType } from './AuthContext';
 
 export const useAuth = (): AuthContextType => {
   const context = useContext(AuthContext);
