--- src/types/assets.d.ts
@@ -1,4 +1,0 @@
-declare module '*.png' {
-  const content: string;
-  export default content;
-}
